rf: Code cleanup (files and imports); move functions between files
Code cleanup (files and imports).
Move functions between files based on the scope of the file.
