+++ src/index.ts
@@ -1,4 +1,4 @@
 export * from "./lib/types";
 export * from "./lib/fastDownward/fdTypes";
 export * from "./lib/pathPlanner";
-export * from "./lib/fastDownward/missingSkillDetection";
+export * from "./lib/backward-search/backward-search";
