+++ src/lib/fastDownward/search-node-list.ts
@@ -1,5 +1,5 @@
 import { LearningUnit, Unit } from "../types";
-import { SearchNode } from "./searchNode-new";
+import { SearchNode } from "./searchNode";
 import { State } from "./state";
 
 // Bucket type to store the search nodes and facilitate their replacement
